Chapter09/requirements.txt
absl-py==0.3.0
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.4.16
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
graphviz==0.8
grpcio==1.12.1
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
mkl-fft==1.0.4
mkl-random==1.0.1
mock==2.0.0
mxnet==0.11.0rc1
numpy==1.14.5
olefile==0.45.1
pandas==0.23.2
pbr==4.2.0
Pillow==5.2.0
protobuf==3.5.2
pyparsing==2.2.0
pyro-ppl==0.2.0+9c3d68e
python-dateutil==2.7.3
pytz==2018.5
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.2
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
tornado==5.1
tqdm==4.23.4
Werkzeug==0.14.1
wget==3.2
